feat(electron): minify renderer and replace mermaid.js with beautiful-mermaid
Enable renderer JS/CSS minify, isolate beautiful-mermaid/elkjs into mermaid-deps, alias shiki/bundle/full to shiki, and add a hash-gated markdown fixture for desktop verification.

Model: grok-4.6
