+++ packages/components/package.json
@@ -124,6 +124,7 @@
     "@xterm/addon-fit": "^0.10.0",
     "@xterm/xterm": "^5.5.0",
     "anser": "^2.3.3",
+    "beautiful-mermaid": "^1.1.3",
     "better-auth": "catalog:",
     "better-auth-capacitor": "catalog:",
     "broadcast-channel": "^7.0.0",
