+++ packages/components/src/components/ai-gui/markdown-renderer.tsx
@@ -10,7 +10,6 @@ import {
   memo,
 } from 'react';
 import { createMathPlugin } from '@streamdown/math';
-import type { MermaidConfig } from 'mermaid';
 import rehypeRaw from 'rehype-raw';
 import rehypeSanitize from 'rehype-sanitize';
 import remarkGfm from 'remark-gfm';
@@ -20,7 +19,6 @@ import {
   type CodeHighlighterPlugin,
   type Components,
   type ControlsConfig,
-  type DiagramPlugin,
   type HighlightOptions,
   type MermaidOptions,
   type PluginConfig,
@@ -50,9 +48,12 @@ import {
   useSessionSearchBlock,
 } from '@/components/sessions/session-search-context';
 import { findSessionSearchOccurrences } from '@/lib/session-chat-search';
-import { useResolvedTheme, type ResolvedTheme } from '../../theme-provider';
+import { useResolvedTheme } from '../../theme-provider';
 import type { ConversationFontSize } from '@/atoms/settings';
 import { useTaskImageUrl } from '@/hooks/use-task-image';
+import { createMarkdownMermaidConfig, createMarkdownMermaidPlugin } from './markdown-mermaid';
+
+export { createMarkdownMermaidConfig } from './markdown-mermaid';
 
 type MarkdownCodeProps = ComponentPropsWithoutRef<'code'> & {
   inline?: boolean;
@@ -540,62 +541,27 @@ const createLazyShikiCodePlugin = (): CodeHighlighterPlugin => {
   let highlighterPromise: Promise<ShikiHighlighter> | null = null;
 
   const loadHighlighter = async () => {
-    // Use core + explicit grammar imports instead of import('shiki');
-    // the bundled entry exposes every Shiki language/theme to Vite chunking.
+    // @pierre/diffs already imports shiki's bundledLanguages catalog. Reuse it
+    // instead of a second shiki/langs/*.mjs graph (duplicate grammar chunks).
     highlighterPromise ??= Promise.all([
       import('shiki/core'),
       import('shiki/engine/javascript'),
-      import('shiki/langs/typescript.mjs'),
-      import('shiki/langs/tsx.mjs'),
-      import('shiki/langs/javascript.mjs'),
-      import('shiki/langs/jsx.mjs'),
-      import('shiki/langs/json.mjs'),
-      import('shiki/langs/shellscript.mjs'),
-      import('shiki/langs/markdown.mjs'),
-      import('shiki/langs/diff.mjs'),
-      import('shiki/langs/python.mjs'),
-      import('shiki/langs/rust.mjs'),
-      import('shiki/langs/go.mjs'),
-      import('shiki/langs/yaml.mjs'),
-      import('shiki/langs/html.mjs'),
-      import('shiki/langs/css.mjs'),
+      import('shiki'),
     ]).then(
       ([
         { createCssVariablesTheme, createHighlighterCore },
         { createJavaScriptRegexEngine },
-        typescript,
-        tsx,
-        javascript,
-        jsx,
-        json,
-        shellscript,
-        markdown,
-        diff,
-        python,
-        rust,
-        go,
-        yaml,
-        html,
-        css,
+        shiki,
       ]) =>
         createHighlighterCore({
           engine: createJavaScriptRegexEngine(),
-          langs: [
-            typescript.default,
-            tsx.default,
-            javascript.default,
-            jsx.default,
-            json.default,
-            shellscript.default,
-            markdown.default,
-            diff.default,
-            python.default,
-            rust.default,
-            go.default,
-            yaml.default,
-            html.default,
-            css.default,
-          ],
+          langs: MARKDOWN_CODE_LANGUAGES.map((id) => {
+            const language = shiki.bundledLanguages[id];
+            if (!language) {
+              throw new Error(`Missing bundled shiki language: ${id}`);
+            }
+            return language;
+          }),
           themes: [
             createCssVariablesTheme({
               name: MARKDOWN_CODE_THEME_NAME,
@@ -637,213 +603,7 @@ const createLazyShikiCodePlugin = (): CodeHighlighterPlugin => {
 
 const MARKDOWN_CODE_PLUGIN = createLazyShikiCodePlugin();
 
-type MermaidRuntime = (typeof import('mermaid'))['default'];
-
-const MERMAID_FONT_FAMILY =
-  'Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", sans-serif';
-
-const MERMAID_NATURAL_WIDTH_CONFIG = {
-  architecture: { useMaxWidth: false },
-  block: { useMaxWidth: false },
-  c4: { useMaxWidth: false },
-  class: { useMaxWidth: false },
-  er: { useMaxWidth: false },
-  flowchart: { useMaxWidth: false },
-  gantt: { useMaxWidth: false },
-  gitGraph: { useMaxWidth: false },
-  journey: { useMaxWidth: false },
-  kanban: { useMaxWidth: false },
-  mindmap: { useMaxWidth: false },
-  packet: { useMaxWidth: false },
-  pie: { useMaxWidth: false },
-  quadrantChart: { useMaxWidth: false },
-  requirement: { useMaxWidth: false },
-  sequence: { useMaxWidth: false },
-  state: { useMaxWidth: false },
-  timeline: { useMaxWidth: false },
-  xyChart: { useMaxWidth: false },
-} satisfies Partial<MermaidConfig>;
-
-const MERMAID_BASE_CONFIG = {
-  ...MERMAID_NATURAL_WIDTH_CONFIG,
-  fontFamily: MERMAID_FONT_FAMILY,
-  securityLevel: 'strict',
-  startOnLoad: false,
-  suppressErrorRendering: true,
-  theme: 'base',
-} satisfies MermaidConfig;
-
-const MERMAID_LIGHT_THEME_VARIABLES = {
-  background: '#ffffff',
-  mainBkg: '#f8fafc',
-  secondaryColor: '#eef6ff',
-  tertiaryColor: '#f1f5f9',
-  primaryColor: '#eef6ff',
-  primaryBorderColor: '#60a5fa',
-  primaryTextColor: '#0f172a',
-  secondaryTextColor: '#0f172a',
-  tertiaryTextColor: '#0f172a',
-  lineColor: '#64748b',
-  textColor: '#0f172a',
-  titleColor: '#0f172a',
-  defaultLinkColor: '#64748b',
-  edgeLabelBackground: '#ffffff',
-  nodeBorder: '#60a5fa',
-  clusterBkg: '#f8fafc',
-  clusterBorder: '#cbd5e1',
-  actorBkg: '#f8fafc',
-  actorBorder: '#60a5fa',
-  actorTextColor: '#0f172a',
-  signalColor: '#64748b',
-  signalTextColor: '#0f172a',
-  labelBoxBkgColor: '#ffffff',
-  labelBoxBorderColor: '#cbd5e1',
-  labelTextColor: '#0f172a',
-  loopTextColor: '#0f172a',
-  noteBkgColor: '#fffbeb',
-  noteTextColor: '#78350f',
-  noteBorderColor: '#f59e0b',
-  activationBkgColor: '#dbeafe',
-  activationBorderColor: '#60a5fa',
-  sequenceNumberColor: '#475569',
-};
-
-const MERMAID_DARK_THEME_VARIABLES = {
-  background: '#0b1120',
-  mainBkg: '#111827',
-  secondaryColor: '#172033',
-  tertiaryColor: '#1e293b',
-  primaryColor: '#172033',
-  primaryBorderColor: '#60a5fa',
-  primaryTextColor: '#f8fafc',
-  secondaryTextColor: '#f8fafc',
-  tertiaryTextColor: '#f8fafc',
-  lineColor: '#cbd5e1',
-  textColor: '#e2e8f0',
-  titleColor: '#f8fafc',
-  defaultLinkColor: '#cbd5e1',
-  edgeLabelBackground: '#0b1120',
-  nodeBorder: '#60a5fa',
-  clusterBkg: '#0f172a',
-  clusterBorder: '#475569',
-  actorBkg: '#111827',
-  actorBorder: '#60a5fa',
-  actorTextColor: '#f8fafc',
-  signalColor: '#cbd5e1',
-  signalTextColor: '#f8fafc',
-  labelBoxBkgColor: '#111827',
-  labelBoxBorderColor: '#475569',
-  labelTextColor: '#f8fafc',
-  loopTextColor: '#f8fafc',
-  noteBkgColor: '#422006',
-  noteTextColor: '#fffbeb',
-  noteBorderColor: '#f59e0b',
-  activationBkgColor: '#1e3a5f',
-  activationBorderColor: '#60a5fa',
-  sequenceNumberColor: '#cbd5e1',
-};
-
-export const createMarkdownMermaidConfig = (theme: ResolvedTheme): MermaidConfig => ({
-  ...MERMAID_BASE_CONFIG,
-  darkMode: theme === 'dark',
-  themeVariables:
-    theme === 'dark' ? { ...MERMAID_DARK_THEME_VARIABLES } : { ...MERMAID_LIGHT_THEME_VARIABLES },
-});
-
-const escapeXml = (value: string): string =>
-  value
-    .replaceAll('&', '&amp;')
-    .replaceAll('<', '&lt;')
-    .replaceAll('>', '&gt;')
-    .replaceAll('"', '&quot;')
-    .replaceAll("'", '&#39;');
-
-// Mermaid's runtime ships regex constructs (e.g. lookbehind) that older Safari
-// (iOS < 16.4) cannot parse, throwing SyntaxError at chunk load time. If we let
-// that propagate, the whole app boot can die because mermaid's chunk sits next
-// to code paths the router reaches. Render a static SVG fallback instead so
-// the rest of the UI keeps working.
-const renderMermaidLoadFallback = (source: string): { svg: string } => {
-  const lines = source.split('\n').slice(0, 12);
-  const truncated = source.split('\n').length > lines.length;
-  const text = lines.map((line) => escapeXml(line)).join('\n');
-  const svg = `<svg xmlns="http://www.w3.org/2000/svg" width="480" height="${
-    24 + lines.length * 16 + (truncated ? 16 : 0)
-  }" viewBox="0 0 480 ${24 + lines.length * 16 + (truncated ? 16 : 0)}">
-  <rect width="100%" height="100%" fill="#fef3c7" rx="6" />
-  <text x="12" y="18" font-family="ui-monospace,monospace" font-size="11" fill="#92400e">Diagram unavailable in this browser</text>
-  <text x="12" y="40" font-family="ui-monospace,monospace" font-size="11" fill="#78350f"><tspan>${text}</tspan>${
-    truncated ? '<tspan x="12" dy="16">…</tspan>' : ''
-  }</text>
-</svg>`;
-  return { svg };
-};
-
-const createLazyMermaidPlugin = (): DiagramPlugin => {
-  let runtimePromise: Promise<MermaidRuntime> | null = null;
-  let runtimeUnavailable = false;
-  let currentConfig: MermaidConfig = createMarkdownMermaidConfig('light');
-
-  const loadRuntime = async (): Promise<MermaidRuntime | null> => {
-    if (runtimeUnavailable) return null;
-    runtimePromise ??= import('mermaid').then((module) => module.default);
-    try {
-      const runtime = await runtimePromise;
-      runtime.initialize(currentConfig);
-      return runtime;
-    } catch (error) {
-      runtimeUnavailable = true;
-      runtimePromise = null;
-      console.warn(
-        '[Lody] Mermaid runtime failed to load; falling back to static text. ' +
-          'This usually means the browser is older than Safari 16.4 or Chrome 88.',
-        error
-      );
-      return null;
-    }
-  };
-
-  const instance = {
-    initialize: (config: MermaidConfig) => {
-      currentConfig = { ...MERMAID_BASE_CONFIG, ...config };
-      if (runtimePromise) {
-        void runtimePromise
-          .then((runtime) => {
-            runtime.initialize(currentConfig);
-          })
-          .catch(() => {
-            // Already handled in loadRuntime; nothing to do here.
-          });
-      }
-    },
-    render: async (id: string, source: string) => {
-      const runtime = await loadRuntime();
-      if (!runtime) {
-        return renderMermaidLoadFallback(source);
-      }
-      try {
-        return await runtime.render(id, source);
-      } catch (error) {
-        console.warn('[Lody] Mermaid render failed; using fallback.', error);
-        return renderMermaidLoadFallback(source);
-      }
-    },
-  };
-
-  return {
-    name: 'mermaid',
-    type: 'diagram',
-    language: 'mermaid',
-    getMermaid: (config?: MermaidConfig) => {
-      if (config) {
-        instance.initialize(config);
-      }
-      return instance;
-    },
-  };
-};
-
-const MARKDOWN_MERMAID_PLUGIN = createLazyMermaidPlugin();
+const MARKDOWN_MERMAID_PLUGIN = createMarkdownMermaidPlugin();
 
 const STREAMDOWN_PLUGINS = {
   code: MARKDOWN_CODE_PLUGIN,
